moved common functions of SapiEmitter to a trait #107
